création de la class scrapper et ajout du transcript avec chapitrage
